finish refactoring map backend into websocket package
