--- mirror-map/src/main/java/mirrormap/mapbackend/ClientHandler.java
@@ -1,56 +1,0 @@
-package mirrormap.mapbackend;
-
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.io.PrintWriter;
-import java.net.Socket;
-
-public class ClientHandler implements Runnable {
-    private Socket clientSocket;
-    private PrintWriter out;
-    private BufferedReader in;
-
-    public ClientHandler(Socket socket) {
-        this.clientSocket = socket;
-        try {
-            out = new PrintWriter(clientSocket.getOutputStream(), true);
-            in = new BufferedReader(new InputStreamReader(clientSocket.getInputStream()));
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-
-    @Override
-    public void run() {
-        String inputLine;
-        try {
-            while ((inputLine = in.readLine()) != null) {
-                System.out.println("Received: " + inputLine);
-                broadcast(inputLine);
-            }
-        } catch (IOException e) {
-            e.printStackTrace();
-        } finally {
-            try {
-                in.close();
-                out.close();
-                clientSocket.close();
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-        }
-    }
-
-    private void broadcast(String message) {
-        for (ClientHandler client : MapBackend.getClients()) {
-            if (client != this) {
-                client.sendMessage(message);
-            }
-        }
-    }
-
-    private void sendMessage(String message) {
-        out.println(message);
-    }
-}
