--- mirror-map/src/main/java/mirrormap/mapbackend/MapBackend.java
@@ -1,31 +1,0 @@
-package mirrormap.mapbackend;
-
-import java.io.IOException;
-import java.net.ServerSocket;
-import java.net.Socket;
-import java.util.ArrayList;
-import java.util.List;
-
-public class MapBackend {
-    private static final int PORT = 8081;
-    private static List<ClientHandler> clients = new ArrayList<>();
-
-    public static void main(String[] args) {
-        try (ServerSocket serverSocket = new ServerSocket(PORT)) {
-            System.out.println("Map Backend listening on port " + PORT);
-            while (true) {
-                Socket clientSocket = serverSocket.accept();
-                System.out.println("New client connected: " + clientSocket);
-                ClientHandler clientHandler = new ClientHandler(clientSocket);
-                clients.add(clientHandler);
-                new Thread(clientHandler).start();
-            }
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-
-    public static List<ClientHandler> getClients() {
-        return clients;
-    }
-}
