--- mirror-map/src/main/java/mirrormap/mapbackend/MapBackendDriver.java
@@ -1,7 +1,0 @@
-package mirrormap.mapbackend;
-
-public class MapBackendDriver {
-    public static void main(String[] args) {
-        MapBackend.main(args);
-    }
-}
